rename: components/common dir 생성 및 공통 컴포넌트 이동, import  경로 수정
