+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import Header from "../components/Header";
+import Header from "../components/common/Header";
 
 export default function Home() {
   return (
